fix: eslint and prettier things

+++ services/brand.js
@@ -1,38 +1,46 @@
-const brandModel = require("../models/brand");
+const { missingbrandData } = require('../constants/serviceErrors');
+const brandModel = require('../models/brand');
 
-async function findAll(filter = null) {
-    try {
-        return await brandModel.find(filter);
-    } catch (error) {
-        console.log(error);
-    }
-}
+// Obtener todos los brands
+const findAll = async function findAll(filter = null) {
+  try {
+    return await brandModel.find(filter);
+  } catch (error) {
+    return error;
+  }
+};
 
-async function create(brandData) {
-    try {
-        return await brandModel.create(brandData);
-    } catch (error) {
-        console.log(error);
-    }
-}
+// Crea un Brand
+const create = async function create(brandData) {
+  if (!brandData) throw missingbrandData;
+  try {
+    return await brandModel.create(brandData);
+  } catch (error) {
+    return error;
+  }
+};
 
-async function deleteById(brandId) {
-    try {
-        let brandDeleted = await brandModel.deleteOne({ _id: brandId });
-        return brandDeleted;
-    } catch (error) {
-        throw error;
-    }
-}
+// Elimina un brand por ID
+const deleteById = async function deleteById(brandId) {
+  if (!brandId) throw missingbrandData;
+  try {
+    const brandDeleted = await brandModel.deleteOne({ _id: brandId });
+    return brandDeleted;
+  } catch (error) {
+    return error;
+  }
+};
 
-async function updateById(brandId, brandData) {
-    try {
-        let brandUpdate = await brandModel.updateOne({ _id: brandId }, brandData);
-        return brandUpdate;
-    }
-    catch (error) {
-        throw error;
-    }
-}
+// Actualiza info de Brand
+const updateById = async function updateById(brandId, brandData) {
+  if (!brandId) throw missingbrandData;
+  if (!brandData) throw missingbrandData;
+  try {
+    const brandUpdate = await brandModel.updateOne({ _id: brandId }, brandData);
+    return brandUpdate;
+  } catch (error) {
+    return error;
+  }
+};
 
-module.exports = { findAll, create, deleteById, updateById, };
+module.exports = { findAll, create, deleteById, updateById };
